tools/requirements.txt
torch==2.2.1
numpy==1.26.4
pandas==2.2.1
transformers==4.38.1
scikit-learn==1.4.0
openpyxl==3.1.2
sentence-transformers==2.4.0
seaborn==0.13.2
matplotlib==3.8.3
plotly==5.19.0
openai==1.13.3
nltk==3.8.1
typing==3.7.4.3
ctransformers==0.2.27
anyascii==0.3.2
python-docx==1.1.0
huggingface-hub==0.21.4
nose2==0.14.1
junit2html
nose2
coverage
